Rename stream_serializer.h
