+++ source/library/reflex/public/potato/reflex/CMakeLists.txt
@@ -4,6 +4,6 @@ target_sources(reflex PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/metadata.h
     ${CMAKE_CURRENT_SOURCE_DIR}/reflect.h
     ${CMAKE_CURRENT_SOURCE_DIR}/json_serializer.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/stream_serializer.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/serializer.h
     ${CMAKE_CURRENT_SOURCE_DIR}/traits.h
 )
