+++ source/library/reflex/tests/test_serialize.cpp
@@ -1,5 +1,5 @@
 #include "potato/reflex/reflect.h"
-#include "potato/reflex/stream_serializer.h"
+#include "potato/reflex/serializer.h"
 #include "potato/reflex/json_serializer.h"
 #include <doctest/doctest.h>
 #include <sstream>
